refactor(DB): separated class models for DB
