+++ step1-clean/db.py
@@ -1,23 +1,8 @@
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import create_engine, Column, String, Integer
+from sqlalchemy import create_engine
+import models
 
-Base = declarative_base()
-
-class SQLImage(Base):
-  __tablename__ = 'Thumbnails'
-
-  id = Column(Integer, primary_key=True)
-  filename = Column(String)
-  isThumbnail = Column(String)
-
-  def __repr__(self):
-    return "SQLImage<id={id}, filename='{filename}' isThumbnail='{isThumbnail}'>".format(
-      id=self.id,
-      filename=self.filename,
-      isThumbnail=self.isThumbnail
-    )
-
+Base = models.Base
 
 engine = create_engine('sqlite+pysqlite:///GG.db')
 Base.metadata.create_all(engine)
